Improve reverse 2FA mode
Adds a new listener to expand XenForo_Session. This is neccessary as
otherwise XenForo always checks the IP of the session, which does of course
never equal the IP of the Threema Gateway.

Some FAQ entries for the callback exceptions and logging were added, so
server admins are not surprised, because of this exceptions.