For deleted messages the receive date is now stil saved in a rounded way
The previous way with deleting all metadata meant that one cannot say when
the message has been send/received (which is intended). However this also
means that a cron job cannot delete these entries later as it cannot deter-
minate whether an entry is old enough to be deleted. (older than 14 days
so it cannot be used for a replay attack detection)

Also some other small cleanup were made.