Fix security issue where Fast 2FA mode could be circumvented
'receivedDeliveryReceiptLargest' always was the largest value, which is usually
"confirm/acknowledge" and as message ID & co are always updated even when the
delivery receipt is just a read or receive confirmation. Thus an attacker
could just click on the confirm button without the need to verify the message
if the user logged in at least once before with this 2FA mode.

To prevent this the data is now always reset after a successful verification,
also for all other 2FA modes to prevent similar issues like this one.