2FA Fast mode: Let Gateway callback trigger the block/decline code
If a user delines a message this now immediately triggers the blocking actions instead
of having to wait until the attacker clicks on the confirm button or so.
Additionally the default expiration/time limit was lifted, so that users can also
decline a message up to 1h later to block a user.