Modify 2FA attempt limit if needed
This commit finalizes the automatic validation trigger.
Now the login actually works as I rose the attempt limit for the 2FA modes, so that one request
per 2 seconds is allowed.
Additionally some templates were modified, so the login also works in the ACP and the
(manualy & optional) fingerprint verification by the user is done before acknowledging
the message in the fast mode.