😢 Goodbye project… I'll freeze you.
It was a great project to learn and fun to made, but I actually never
heard of anyone using it. That's no problem at all, it must not be used by
anyone, but I also could not afford the time to adjust it to the new XenForo
version, so when I started the development I had quite a bad timing.
Nevertheless may the code be useful to someone, it will stay here.

May the code rest in peace.